+++ infrastructure-provisioning/cloud/pulumi/index.ts
@@ -2,8 +2,9 @@ import * as pulumi from "@pulumi/pulumi";
 import * as aws from "@pulumi/aws";
 import * as awsx from "@pulumi/awsx";
 import * as config from "./vars/environment"
-import { getUbuntuAmi } from "./utils/getAmi"
 import { InstanceType } from "@pulumi/aws/ec2";
+// Imports
+import { createK3sWorkers } from "./utils/k3sWorkers"
 
 const vpc = new aws.ec2.Vpc("primary", {
   cidrBlock: config.network.vpc.cidr_block,
@@ -17,22 +18,11 @@ for (const subnet of config.network.subnets.public) {
     vpcId: vpc.id,
     cidrBlock: subnet.cidr_block
   })
-  publicSubnets.push({"name": subnet.name, "id": s.id})
+  publicSubnets.push({ "name": subnet.name, "id": s.id })
 }
 
-// Provision a K3s worker
-const k3sworkers = []
-
-for (const node of config.compute.workers) {
-  const s = new aws.ec2.Instance(node.name,{
-    ami: getUbuntuAmi().then(ubuntu => ubuntu.id),
-    instanceType: node.instance_size,
-    rootBlockDevice: {
-      volumeSize: node.root_volume_size,
-    },
-    tags: config.tags,
-  })
-}
+// Create K3s Workers
+const workers = createK3sWorkers(publicSubnets)
 
 // // Export the name of the bucket
-// export { publicSubnets }
+export { createK3sWorkers }
